+++ src/main/java/org/jboss/classfilewriter/constpool/ClassEntry.java
@@ -21,9 +21,9 @@ import java.io.DataOutputStream;
 import java.io.IOException;
 
 public class ClassEntry extends ConstPoolEntry {
-    private final short utf8Location;
+    private final int utf8Location;
 
-    public ClassEntry(short utf8Location) {
+    public ClassEntry(int utf8Location) {
         this.utf8Location = utf8Location;
     }
