+++ src/main/java/org/jboss/classfilewriter/constpool/InterfaceMethodRefEntry.java
@@ -22,10 +22,10 @@ import java.io.IOException;
 
 public class InterfaceMethodRefEntry extends ConstPoolEntry {
 
-    private final short classIndex;
-    private final short nameAndTypeIndex;
+    private final int classIndex;
+    private final int nameAndTypeIndex;
 
-    public InterfaceMethodRefEntry(short classIndex, short nameAndTypeIndex) {
+    public InterfaceMethodRefEntry(int classIndex, int nameAndTypeIndex) {
         this.classIndex = classIndex;
         this.nameAndTypeIndex = nameAndTypeIndex;
     }
