+++ src/main/java/org/jboss/classfilewriter/constpool/StringEntry.java
@@ -22,9 +22,9 @@ import java.io.IOException;
 
 public class StringEntry extends ConstPoolEntry {
 
-    private final short utf8Index;
+    private final int utf8Index;
 
-    public StringEntry(short utf8Index) {
+    public StringEntry(int utf8Index) {
         this.utf8Index = utf8Index;
     }
